+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v4.8.1
 * Added: Term aggregation result includes `sum_other_doc_count` and `doc_count_error_upper_bound`
+* Added: Term aggregation parameter named `includeHasNotCount` to include a count of the number of element missing the aggregated property. This is false by default.
 
 # v4.8.0
 * Added: Option to refresh on flush as opposed to flushing at query time